requirements.txt
alpaca-py
pandas
numpy
joblib
scikit-learn
nest-asyncio
python-dotenv